Update boreholeGeometryLog.xml
This model of a borehole geometry log can also be used as a template for any type of observation in a borehole. It examplarily specifies attributes like constraints, quality information, units of measure, and many more. Of great importance is the specification of vocabularies, by which the various fields can be referenced uniquely.